chore: fix up detection
